[level 3] Title: 디스크 컨트롤러, Time: 3.65 ms, Memory: 85.8 MB -BaekjoonHub

+++ 프로그래머스/3/42627. 디스크 컨트롤러/디스크 컨트롤러.java
@@ -51,7 +51,7 @@ class Solution {
         }
 
         // 소수점 이하는 버림 --> 정수형으로 변환
-        answer = (int)(total / jobs.length);
+        answer = total / jobs.length;
         return answer;
     }
 }
